SDL2, Joystick, UTF8, MemoryLeaks, Fonts, CampaignMessages, Log, and more...
* ------------------------------------------------------------------------
* bug fixes since 1.03 :
  + Oxford Matterson library message was not delivered to the right room, but
    the the launch pad.

  + in addition to the oxford fix, now sounds are also played in the right room

  + python missions database was cleared on success of any mission,
    which lead to an empty mission list in QUINE even if there were others
    (old basecomputer could see them). To reproduce the bug on old
    version, accept 2 missions, finsih one and land. you will see on
    Quine an empty mission list. Cause: all missions had same ID=0.

  + probably fixed: random crash when jumping : i did reproduce it
    with gdb (crash in gl_light_state>unpicklights(), then applied a patch provided
    by a 2012 svn commit (see rev 297), and did not see the issue again after hours,
    of jumps and starsystems garbage collection.
    In addition i changed something in universe->StartDraw() to prevent a system
    to be garbage collected before the python scripts have run.

  + overwrite savegame in QUINE, now confirmation is handled well even if the
    existing file is typed by user. And 'New_Game' name is forbidden.

  + QUINE: SaveGames with underscores are correctly displayed.

  + time was running too fast when several missions were on-going
    (increment of SIMULATION_ATOM per mission)

* ------------------------------------------------------------------------
* NOT fixed since 1.03:
  + addition of random parts to ship, the only way to solve it
    for now is to edit the '~/.privgold100/serialized_xml/<save>/<ship>.begin.csv,
    and remove the part {...}.
  + still some memory leaks, but much less. The game can run for hours.
    Part of the Memory usage is normal as the game caches 15 star systems, but some
    other are not. I started easyDomNode cleaning, and finished-(Mission*) cleaning,
    but of course it is not enough/finished.
  + in some campaign dialogs with accept/reject option, the accept/reject is not
    cleared if we leave the room without answering

* ------------------------------------------------------------------------
* engine: override a XML config variable via command line
  (eg: ./vegastrike -Cgraphics/fullscreen=no)

* engine/python: LOGGING
  + starting to use a common log module,
    progressively replacing printfs, cout, ...

* the cockpit is no more shown during game loading before to show the base

* the Navigation Screen margins computation for the System and Sector Map
  has been improved for 16/9 resolutions (1680x1050,1280x800,...).
  No Change for 4/3 resolutions

* message center alternative to handle fixers messages during campaign battles
  (when a campaign fixer has some thing to tell you, it will only
  displayed when docked at a base, now, it is display during flight.

* engine: HashTable and Unit Container
  + don't use tr1 unordered maps
  + use gnu hash_map or cxx11 unordered_map
  + use New STL collection or old collection.
    Currently Old one is used.

* engine: SDL2 support
  SDL2 provides better unicode support, better window management support (resize, ...),
  and joystick plug and play support.
  - engine/sdl2: alt+tab will minimize game in fullscreen
  - SDL: possibility to disable the SDL LockAudio workaround

* Joystick/Mouse: reduce polling, joystick when docked
  + Only SDL is polling the joystick (via SDL_PollEvent).
    Now the mouse and joystick interface register to a joystick
    winsys event instead of polling joystick state.
  + With SDL2, the joystick can be plugged/unplugged while the
    the game is running.

* BaseInterface Links&Objects lookup optimization and improvement of memory leaks
  + Links&Objects lookup optimization
    Now Links and Object are stored in a multimap which speeds up the
    lookups and deletion. Because the order of insertion matters when
    drawing, the multimap contains chained object in order of insertion
    without adding complexity to multimap lookup/insertion/deletion.
    This redesign allowed to fix a huge memory leak (can double the RAM
    usage in few minutes).
  + Links Navigation: we can now navigate between links with keyboard
    (arrow keys, space, enter) or joystick

* engine: improved sdl1 and glut support:
  + sdl1/keyboard: alt+x or any alt+ascii were not working on azerty mac,
    and maybe qwerty also, because of alt+key on mac gives unicode.
  + glut/keyboard: handle utf8 input
  + glut/mousewheel: emulation of mouse wheel with ctrl+up/down on glut
    (no wheel support on glut for mac)

* engine/data: unicode(utf8) support on keyboard and on display
  + engine/data: update all all gui elements taking keyboard input
  + engine/data: update some GUI elements that could display unicode
  + engine/data: management '##' escape sequence used to display a '#'
  + engine: update of Font class
  + engine: glut/sdlx updates to handle unicode 32bits input
  + Still TODO: the gfx/gui PaintText class is not fully utf8.
    The Trucation of text needs some more work. Not done because, on
    Privateer, the PaintText class is only used by the 'OLD Base Computer'

* engine/setup:
  + build both gtk and dialog versions
  + gtk/dialog: if available, open Manual.pdf instead of readme
  + choose fonts, graphic engine, unicode configuration, setup program.

* bundle:
  + regroup game and setup in the same bundle
  + run Setup by pressing 'alt' when running the game
  + add 3 differents graphical engines (separate binaries): GLUT, SDL1, SDL2.

* main_loop:
  + optional display of FPS and loop rate (alt+p)
  + optional display of key bindings during flight (alt+h)
  + optional display of logs during flight via message center (alt+m)

* data: UTF8 Savegames AND home folder
  + home folder is now .privgold120. When running the app bundle,
    the .privgold100 is automatically duplicated to .privgold120
  + This version is backward-campatible with 1.0x (.privGold100).
    But the contrary is not fully true about savegames. Indeed,
    because of UTF8 support, the savegames saved from new version
    are not directly compatible with old one, but the tool
    unicode-conv (Contents/Resources/bin/unicode-conv) can convert a new
    savegame to the 1.0x old format so that they run on old version:
    'unicode-conv --convert-back ~/.privgold120/save/Game old_Game'
  + When loading an old savegame, it is converted on-the-fly to utf8.

* tools
  + build objconv tools and others with cmake
  + add unitary tests for Utf8Iterator and ChainedMultimap
  + add a tool converting a file or privateer savegame to/from UTF8
  + Generation of the C++ Python API files (empty functions) in order
    to be used in an IDE Completion System
  + Detection of Space/Tabs usage in python files

+++ vegastrike/src/hashtable.h
@@ -38,7 +38,7 @@ template<class KEY, class VALUE, int SIZ> class Hashtable :public vsUMap <KEY,VA
   typedef std::pair<KEY,VALUE*> HashElement;
   typedef vsUMap<KEY,VALUE* > supertype;
 public:
-	static int hash(const int key) {
+	static size_t hash(const int key) {
 		unsigned int k = key;
 		k%=SIZ;
 		return k;
@@ -54,7 +54,7 @@ public:
           k %= SIZ;
           return k;
 	}
-	static int hash(const std::string &key) {
+	static size_t hash(const std::string &key) {
           unsigned int k = 0;
           for(typename std::string::const_iterator start = key.begin(); start!=key.end(); ++start) {
             k ^= (*start&HASH_SALT_1);
